requirements.txt
flask==2.3.3
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.2